made more secure sharing authentication informations with HTTP and HTTPS in mobile_frontend (refs #1112)
There is no way to realize SSL providable member/setSid action. It is not avoidable an attacker to eavesdrop a connection that contains a session ID, so it permits to being possible session-hijacking.
These changes are for encrypting that session ID in a query string, it makes difficult to imperson a victim