removed the unused method, opWebRequest::hasMobileUidCookie() (fixes #1582)
Do you have an objection to this commit? Do you think that this is BC break? I think not.
First, this method is NOT in stable release. Next, this method is not really useful.
Last, nobody uses this method now, so it should be removed for conforming to YAGNI(http://en.wikipedia.org/wiki/You_ain't_gonna_need_it).