source code tree of stable-3.6.x is now OpenPNE-3.6RC1-dev
Are you surprised by this commit? Adding this commit doesn't mean we will release RC1 soon.
This commit is just for avoiding conflict with source code managiment rule and plan to release 3.6.0 (http://www.openpne.jp/archives/6058/). Do not worry about this commit.

We will release beta{N} from release-3.6beta{N} branch to release well-reviewed and well-tested changes only. Changes in stable-3.6 are not always so now.

Of cource, when we reach 3.6RC1, there are no differences with stable-3.6.x and release-3.6RC1. So current stable-3.6.x is worthy of development version of OpenPNE-3.6RC1.